+++ lib/requirements.php
@@ -21,6 +21,13 @@ require_once WPS_DIR_LIB . '/aihrus/requirements.php';
 
 function wps_requirements_check() {
 	$valid_requirements = true;
+	if ( ! function_exists( 'aihr_check_aihrus_framework' ) ) {
+		$valid_requirements = false;
+		add_action( 'admin_notices', 'wps_notice_aihrus' );
+	} elseif ( ! aihr_check_aihrus_framework( WPS_BASE, WPS_NAME, WPS_AIHR_VERSION ) ) {
+		$valid_requirements = false;
+	}
+
 	if ( ! aihr_check_php( WPS_BASE, WPS_NAME ) ) {
 		$valid_requirements = false;
 	}
@@ -34,6 +41,16 @@ function wps_requirements_check() {
 	}
 
 	return $valid_requirements;
+}
+
+
+function wps_notice_aihrus() {
+	$help_url  = esc_url( 'https://aihrus.zendesk.com/entries/35689458' );
+	$help_link = sprintf( __( '<a href="%1$s">Update plugins</a>. <a href="%2$s">More information</a>.' ), self_admin_url( 'update-core.php' ), $help_url );
+
+	$text = sprintf( esc_html__( 'Plugin "%1$s" has been deactivated as it requires a current Aihrus Framework. Once corrected, "%1$s" can be activated. %2$s' ), WPS_NAME, $help_link );
+
+	aihr_notice_error( $text );
 }
 
 ?>